Implement SplitInstallService
When a 3rd party app wants to install a split APK, the application
tries to bind to a SplitInstallService through the
SplitInstallManager part of the Google Play Feature Delivery
Library.

The implementation of SplitInstallService allows us to catch
the install requests coming from the 3rd party app and handle
them.

As an example, DummyStoreSplitInstaller binds to an external
com.dummy.store.splitinstall.SplitInstallService service which
is able to retrieve the split APKs from the com.dummy.store.